tests: fix most of the rganalysis unit tests
Before the element would post messages on the bus itself, now
the sinks do that based on the tag events they receive. But
since we don't have proper sink elements in these unit tests,
but just dangling pads, we have to post the tag messages the
test checks for ourselves.

Down from 52/55 failing to 7/52 failing.